refactor: modularize core logic
